S3: Allow CrossAccount access to be disabled
